% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWSI.R
\name{CWSI}
\alias{CWSI}
\title{R6 class for calculating the Crop Water Stress Index (CWSI)}
\description{
destination description This class allows users to load canopy temperature data and micrometeorological data, and then calculate the CWSI based on these inputs.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NA-load_canopyTemp}{\code{CWSI$load_canopyTemp()}}
\item \href{#method-NA-load_micrometeo}{\code{CWSI$load_micrometeo()}}
\item \href{#method-NA-calc_CWSI}{\code{CWSI$calc_CWSI()}}
\item \href{#method-unknown-clone}{\code{CWSI$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-load_canopyTemp"></a>}}
\if{latex}{\out{\hypertarget{method-NA-load_canopyTemp}{}}}
\subsection{Method \code{load_canopyTemp()}}{
Load canopy temperature data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CWSI$load_canopyTemp(Tc, time_series = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Tc}}{Canopy temperature data (°C).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The CWSI object with loaded canopy temperature data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-load_micrometeo"></a>}}
\if{latex}{\out{\hypertarget{method-NA-load_micrometeo}{}}}
\subsection{Method \code{load_micrometeo()}}{
Load micrometeorological data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CWSI$load_micrometeo(
  Rs = NULL,
  Ta = NULL,
  u = NULL,
  rH = NULL,
  h = NULL,
  z = NULL,
  data = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Rs}}{Incoming solar radiation (W/m^2)}

\item{\code{Ta}}{Air temperature (°C)}

\item{\code{u}}{Wind speed (m/s)}

\item{\code{rH}}{Relative humidity (\%)}

\item{\code{h}}{Crop height (m)}

\item{\code{z}}{Measurement height (m)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The CWSI object with loaded micrometeorological data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-calc_CWSI"></a>}}
\if{latex}{\out{\hypertarget{method-NA-calc_CWSI}{}}}
\subsection{Method \code{calc_CWSI()}}{
Calculate the Crop Water Stress Index (CWSI) based on the loaded canopy temperature and micrometeorological data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CWSI$calc_CWSI(method = "theoretical")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{The method to calculate CWSI. Currently, only "theoretical" is implemented.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\deqn{
  T_c - T_A =
  \frac{r_a R_n}{\rho c_p} \cdot \frac{\gamma}{\Delta + \gamma}
  - \frac{(e_A^* - e_A)}{\Delta + \gamma}
}
}

\subsection{Returns}{
A numeric vector of CWSI values for each time point.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unknown-clone"></a>}}
\if{latex}{\out{\hypertarget{method-unknown-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CWSI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
